requirements.txt
aws-cdk-lib==2.166.0
constructs>=10.0.0,<11.0.0
cdklabs.generative-ai-cdk-constructs==0.1.288
aws-cdk.aws-lambda-python-alpha==2.149.0-alpha.0
opensearch-py==2.8.0
boto3==1.35.98
botocore==1.35.98